+++ tsup.config.ts
@@ -2,6 +2,7 @@ import { defineConfig } from "tsup";
 
 export default defineConfig({
   entry: {
+    cli: "src/cli/main.ts",
     main: "src/main/main.ts",
     preload: "src/main/preload.ts",
   },